Add files via upload
Accounting project, version 23A (23B has a slight difference in the mechanics of the 4th frame). Will upload versions 1-22 if I have time and if people want to see the progression. As far as I know, all the versions are functional to the extent to which they are completed. The code itself is sound and works generally as needed. I moved on to a new version if I needed to incorporate some new major element and did not want to retroactively rewrite the iteration on which I was currently working. Version 23 will be the final version, I think.